Split transform classes into separate files
